:sparkles: イベントモデルを新規作成し、記事モデルを削除。イベント関連のAPIエンドポイントを実装
